Move roomba base entity to separate module
