+++ homeassistant/components/roomba/binary_sensor.py
@@ -7,7 +7,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import roomba_reported_state
 from .const import DOMAIN
-from .irobot_base import IRobotEntity
+from .entity import IRobotEntity
 from .models import RoombaData
