+++ homeassistant/components/roomba/entity/braava.py
@@ -4,7 +4,7 @@ import logging
 
 from homeassistant.components.vacuum import VacuumEntityFeature
 
-from .irobot_base import SUPPORT_IROBOT, IRobotVacuum
+from . import SUPPORT_IROBOT, IRobotVacuum
 
 _LOGGER = logging.getLogger(__name__)
